+++ src/components/Navigator.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 import { NavigationContainer } from '@react-navigation/native';
-import Home from './Home/Home';
+import Home from '../Screen/Home';
 
 const Stack = createNativeStackNavigator();
